+++ ModelContextProtocol/McpServer.cs
@@ -11,7 +11,7 @@ namespace TiaMcpServer.ModelContextProtocol
     [McpServerToolType]
     public static class McpServer
     {
-        private static readonly Portal _portal = new Portal();
+        private static readonly Portal _portal = new();
 
         #region portal
 
@@ -65,14 +65,16 @@ namespace TiaMcpServer.ModelContextProtocol
 
         #endregion
 
-        #region project
+        #region project/session
 
-        [McpServerTool, Description("Get list of open projects")]
+        [McpServerTool, Description("Get list of open projects/local sessions")]
         public static string GetOpenProjects()
         {
             try
             {
                 var projects = _portal.GetOpenProjects();
+                projects.AddRange(_portal.GetOpenSessions());
+
                 return JsonRpcMessageWrapper.ToJson(1, projects);
             }
             catch (Exception ex)
@@ -81,7 +83,7 @@ namespace TiaMcpServer.ModelContextProtocol
             }
         }
 
-        [McpServerTool, Description("Open a TIA Portal project/session")]
+        [McpServerTool, Description("Open a TIA Portal project/local session")]
         public static string OpenProject(string projectPath)
         {
             try
@@ -126,23 +128,37 @@ namespace TiaMcpServer.ModelContextProtocol
             }
         }
 
-        [McpServerTool, Description("Save the current TIA Portal project")]
+        [McpServerTool, Description("Save the current TIA Portal project/local session")]
         public static string SaveProject()
         {
             try
             {
-                if (_portal.SaveProject())
+                if (_portal.IsLocalSession)
                 {
-                    return JsonRpcMessageWrapper.ToJson(1, "Project saved successfully.");
+                    if (_portal.SaveSession())
+                    {
+                        return JsonRpcMessageWrapper.ToJson(1, "Local session saved successfully.");
+                    }
+                    else
+                    {
+                        return JsonRpcMessageWrapper.ToJson(1, false, "Failed to save local session.");
+                    }
                 }
-                else
+                else 
                 {
-                    return JsonRpcMessageWrapper.ToJson(1, false, "Failed to save project.");
+                    if (_portal.SaveProject())
+                    {
+                        return JsonRpcMessageWrapper.ToJson(1, "Project saved successfully.");
+                    }
+                    else
+                    {
+                        return JsonRpcMessageWrapper.ToJson(1, false, "Failed to save project.");
+                    }
                 }
             }
             catch (Exception ex)
             {
-                return JsonRpcMessageWrapper.ToJson(1, false, $"Error saving project: {ex.Message}");
+                return JsonRpcMessageWrapper.ToJson(1, false, $"Error saving project/local session: {ex.Message}");
             }
         }
 
@@ -151,18 +167,26 @@ namespace TiaMcpServer.ModelContextProtocol
         {
             try
             {
-                if (_portal.SaveAsProject(newProjectPath))
+                if (_portal.IsLocalSession)
                 {
-                    return JsonRpcMessageWrapper.ToJson(1, $"Project successfully saved as [{newProjectPath}].");
+                    return JsonRpcMessageWrapper.ToJson(1, false, $"Cannot save local session as {newProjectPath}.");
                 }
                 else
                 {
-                    return JsonRpcMessageWrapper.ToJson(1, false, "Failed to save project as.");
+                    if (_portal.SaveAsProject(newProjectPath))
+                    {
+                        return JsonRpcMessageWrapper.ToJson(1, $"Project saved successfully as {newProjectPath}.");
+                    }
+                    else
+                    {
+                        return JsonRpcMessageWrapper.ToJson(1, false, $"Failed to save project as {newProjectPath}.");
+                    }
                 }
+                
             }
             catch (Exception ex)
             {
-                return JsonRpcMessageWrapper.ToJson(1, false, $"Error saving project as: {ex.Message}");
+                return JsonRpcMessageWrapper.ToJson(1, false, $"Error saving project/local session as {newProjectPath}: {ex.Message}");
             }
         }
 
@@ -171,12 +195,23 @@ namespace TiaMcpServer.ModelContextProtocol
         {
             try
             {
-                _portal.CloseProject();
-                return JsonRpcMessageWrapper.ToJson(1, "Project closed successfully.");
+                if (_portal.IsLocalSession)
+                {
+                    _portal.CloseSession();
+
+                    return JsonRpcMessageWrapper.ToJson(1, "Local session closed successfully.");
+                }
+                else
+                {
+                    _portal.CloseProject();
+
+                    return JsonRpcMessageWrapper.ToJson(1, "Project closed successfully.");
+                }
+                
             }
             catch (Exception ex)
             {
-                return JsonRpcMessageWrapper.ToJson(1, false, $"Error closing project: {ex.Message}");
+                return JsonRpcMessageWrapper.ToJson(1, false, $"Error closing project/local session: {ex.Message}");
             }
         }
 
@@ -184,7 +219,20 @@ namespace TiaMcpServer.ModelContextProtocol
 
         #region devices
 
-        [McpServerTool, Description("Get list of devices in the project")]
+        [McpServerTool, Description("Get the structure of the current project/local session")]
+        public static string GetStructure()
+        {
+            try
+            {
+                return _portal.GetStructure();
+            }
+            catch (Exception ex)
+            {
+                return $"Error retrieving structure of the project/local session: {ex.Message}";
+            }
+        }
+
+        [McpServerTool, Description("Get a list of devices in the project")]
         public static List<string> GetDevices()
         {
             try
@@ -193,7 +241,53 @@ namespace TiaMcpServer.ModelContextProtocol
             }
             catch (Exception ex)
             {
-                return new List<string> { $"Error retrieving devices: {ex.Message}" };
+                return [$"Error retrieving devices: {ex.Message}"];
+            }
+        }
+
+        [McpServerTool, Description("Get a device by its path in the project")]
+        public static string GetDevice(string devicePath)
+        {
+            try
+            {
+                var device = _portal.GetDevice(devicePath);
+
+                if (device == null)
+                {
+                    return $"Device '{devicePath}' not found.";
+                }
+
+                return device.Name;
+            }
+            catch (Exception ex)
+            {
+                return $"Error retrieving device: {ex.Message}";
+            }
+        }
+
+        #endregion
+
+        #region plc software
+
+        [McpServerTool, Description("Compile the software, which is given by softwarePath")]
+        public static string CompileSoftware(string softwarePath)
+        {
+            try
+            {
+                var result = _portal.CompileSoftware(softwarePath);
+
+                if (!result.Equals("Error"))
+                {
+                    return JsonRpcMessageWrapper.ToJson(1, $"Software compiled with {result}.");
+                }
+                else
+                {
+                    return JsonRpcMessageWrapper.ToJson(1, false, $"Failed to compile software with {result}");
+                }
+            }
+            catch (Exception ex)
+            {
+                return JsonRpcMessageWrapper.ToJson(1, false, $"Error compiling software: {ex.Message}");
             }
         }
 
@@ -201,12 +295,12 @@ namespace TiaMcpServer.ModelContextProtocol
 
         #region code blocks
 
-        [McpServerTool, Description("Get code block by groupPath/blockName in the project (FB, FC, OB).")]
-        public static string GetCodeBlock(string groupPath, string blockName)
+        [McpServerTool, Description("Get a code block (FB, FC, OB) from the software, which is given by softwarePath/groupPath/blockName.")]
+        public static string GetCodeBlock(string softwarePath, string groupPath, string blockName)
         {
             try
             {
-                return _portal.GetCodeBlock(groupPath, blockName);
+                return _portal.GetCodeBlock(softwarePath, groupPath, blockName);
             }
             catch (Exception ex)
             {
@@ -214,25 +308,25 @@ namespace TiaMcpServer.ModelContextProtocol
             }
         }
 
-        [McpServerTool, Description("Get list of code blocks in the project (FB, FC, OB).")]
-        public static List<string> GetCodeBlocks()
+        [McpServerTool, Description("Get a list of all code blocks (FB, FC, OB) from the software, which is given by softwarePath.")]
+        public static List<string> GetCodeBlocks(string softwarePath)
         {
             try
             {
-                return _portal.GetCodeBlocks();
+                return _portal.GetCodeBlocks(softwarePath);
             }
             catch (Exception ex)
             {
-                return new List<string> { $"Error retrieving code blocks: {ex.Message}" };
+                return [$"Error retrieving code blocks: {ex.Message}"];
             }
         }
 
-        [McpServerTool, Description("Export a code block by groupPath/blockName to a specified exportPath (FB, FC, OB).")]
-        public static string ExportCodeBlock(string groupPath, string blockName, string exportPath)
+        [McpServerTool, Description("Export a code block (FB, FC, OB) given by softwarePath/groupPath/blockName to a specified exportPath.")]
+        public static string ExportCodeBlock(string softwarePath, string groupPath, string blockName, string exportPath)
         {
             try
             {
-                if (_portal.ExportCodeBlock(groupPath, blockName, exportPath))
+                if (_portal.ExportCodeBlock(softwarePath, groupPath, blockName, exportPath))
                 {
                     return JsonRpcMessageWrapper.ToJson(1, "Code block exported successfully.");
                 }
@@ -247,12 +341,12 @@ namespace TiaMcpServer.ModelContextProtocol
             }
         }
 
-        [McpServerTool, Description("Export all code blocks to a specified path (FB, FC, OB).")]
-        public static string ExportCodeBlocks(string exportPath)
+        [McpServerTool, Description("Export all code blocks (FB, FC, OB) from the software, given by softwarePath, to a specified path.")]
+        public static string ExportCodeBlocks(string softwarePath, string exportPath)
         {
             try
             {
-                if (_portal.ExportCodeBlocks(exportPath))
+                if (_portal.ExportCodeBlocks(softwarePath, exportPath))
                 {
                     return JsonRpcMessageWrapper.ToJson(1, "Code blocks exported successfully.");
                 }
@@ -271,12 +365,12 @@ namespace TiaMcpServer.ModelContextProtocol
 
         #region data blocks
 
-        [McpServerTool, Description("Get data block by groupPath/blockName in the project (InstanceDB, GlobalDB, ArrayDB).")]
-        public static string GetDataBlock(string groupPath, string blockName)
+        [McpServerTool, Description("Get a data block (InstanceDB, GlobalDB, ArrayDB), given by softwarePath/groupPath/blockName.")]
+        public static string GetDataBlock(string softwarePath, string groupPath, string blockName)
         {
             try
             {
-                return _portal.GetDataBlock(groupPath, blockName);
+                return _portal.GetDataBlock(softwarePath, groupPath, blockName);
             }
             catch (Exception ex)
             {
@@ -284,25 +378,25 @@ namespace TiaMcpServer.ModelContextProtocol
             }
         }
 
-        [McpServerTool, Description("Get list of data blocks in the project (InstanceDB, GlobalDB, ArrayDB).")]
-        public static List<string> GetDataBlocks()
+        [McpServerTool, Description("Get list of data blocks (InstanceDB, GlobalDB, ArrayDB) from software, given by softwarePath, in the project.")]
+        public static List<string> GetDataBlocks(string softwarePath)
         {
             try
             {
-                return _portal.GetDataBlocks();
+                return _portal.GetDataBlocks(softwarePath);
             }
             catch (Exception ex)
             {
-                return new List<string> { $"Error retrieving data blocks: {ex.Message}" };
+                return [$"Error retrieving data blocks: {ex.Message}"];
             }
         }
 
-        [McpServerTool, Description("Export a data block by groupPath/blockName to a specified exportPath (InstanceDB, GlobalDB, ArrayDB).")]
-        public static string ExportDataBlock(string groupPath, string blockName, string exportPath)
+        [McpServerTool, Description("Export a data block given by softwarePath/groupPath/blockName to a specified exportPath (InstanceDB, GlobalDB, ArrayDB).")]
+        public static string ExportDataBlock(string softwarePath, string groupPath, string blockName, string exportPath)
         {
             try
             {
-                if (_portal.ExportDataBlock(groupPath, blockName, exportPath))
+                if (_portal.ExportDataBlock(softwarePath, groupPath, blockName, exportPath))
                 {
                     return JsonRpcMessageWrapper.ToJson(1, "Data block exported successfully.");
                 }
@@ -317,12 +411,12 @@ namespace TiaMcpServer.ModelContextProtocol
             }
         }
 
-        [McpServerTool, Description("Export all data blocks to a specified path (InstanceDB, GlobalDB, ArrayDB).")]
-        public static string ExportDataBlocks(string exportPath)
+        [McpServerTool, Description("Export all data blocks (InstanceDB, GlobalDB, ArrayDB) from the software, given by softwarePath, to a specified path.")]
+        public static string ExportDataBlocks(string softwarePath, string exportPath)
         {
             try
             {
-                if (_portal.ExportDataBlocks(exportPath))
+                if (_portal.ExportDataBlocks(softwarePath, exportPath))
                 {
                     return JsonRpcMessageWrapper.ToJson(1, "Data blocks exported successfully.");
                 }
@@ -341,12 +435,12 @@ namespace TiaMcpServer.ModelContextProtocol
 
         #region types
 
-        [McpServerTool, Description("Get a user defined type (UDT)")]
-        public static string GetUserDefinedType(string typeName)
+        [McpServerTool, Description("Get a user defined type (UDT) from the software, which is given by softwarePath.")]
+        public static string GetUserDefinedType(string softwarePath, string typeName)
         {
             try
             {
-                return _portal.GetUserDefinedType(typeName);
+                return _portal.GetUserDefinedType(softwarePath, typeName);
             }
             catch (Exception ex)
             {
@@ -354,25 +448,25 @@ namespace TiaMcpServer.ModelContextProtocol
             }
         }
 
-        [McpServerTool, Description("Get list of user defined types (UDT)")]
-        public static List<string> GetUserDefinedTypes()
+        [McpServerTool, Description("Get a list of all user defined types (UDT) from the software, which is given by softwarePath.")]
+        public static List<string> GetUserDefinedTypes(string softwarePath)
         {
             try
             {
-                return _portal.GetUserDefinedTypes();
+                return _portal.GetUserDefinedTypes(softwarePath);
             }
             catch (Exception ex)
             {
-                return new List<string> { $"Error retrieving user defined types: {ex.Message}" };
+                return [$"Error retrieving user defined types: {ex.Message}"];
             }
         }
 
-        [McpServerTool, Description("Export a user defined types (UDT) by name")]
-        public static string ExportUserDefinedType(string exportPath, string typeName)
+        [McpServerTool, Description("Export a user defined types (UDT) by name from the software, which is giveb by softwarePath.")]
+        public static string ExportUserDefinedType(string softwarePath, string exportPath, string typeName)
         {
             try
             {
-                if (_portal.ExportUserDefinedType(exportPath, typeName))
+                if (_portal.ExportUserDefinedType(softwarePath, exportPath, typeName))
                 {
                     return JsonRpcMessageWrapper.ToJson(1, "User defined type exported successfully.");
                 }
@@ -387,12 +481,12 @@ namespace TiaMcpServer.ModelContextProtocol
             }
         }
 
-        [McpServerTool, Description("Export all user defined types (UDT)")]
-        public static string ExportUserDefinedTypes(string exportPath)
+        [McpServerTool, Description("Export all user defined types (UDT) from the software, which is giveb by softwarePath.")]
+        public static string ExportUserDefinedTypes(string softwarePath, string exportPath)
         {
             try
             {
-                if (_portal.ExportUserDefinedTypes(exportPath))
+                if (_portal.ExportUserDefinedTypes(softwarePath, exportPath))
                 {
                     return JsonRpcMessageWrapper.ToJson(1, "User defined types exported successfully.");
                 }